Patch campilation error by making context members usize
It seems at some point snowball_env.rs has been updated in the rust
snowball backend but the change is not reflected in rust-stemmers
project.

The right fix probably is to fix snowball_env.rs and regenerate all rust
code for language-specific stemmers. But this small change at least
makes the main branch compile out of the box again.

See #20 for an additional github action suggestion that runs tests on
PRs